requirements.txt
asgiref==3.3.1
Django==3.1.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
mysqlclient==2.0.3
Pillow==8.1.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
